+++ projectify/settings/development.py
@@ -89,16 +89,15 @@ class Development(Base):
     EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
     SERVER_EMAIL = "hello@projectifyapp.com"
 
-    # Media
+    # Static files
+    STATIC_ROOT = Base.BASE_DIR / "staticfiles"
+
+    # Media files
+    MEDIA_ROOT = Base.BASE_DIR / "media"
     SENDFILE_BACKEND = "django_sendfile.backends.simple"
 
     # Safari workaround for sessionid cookie
     SESSION_COOKIE_SECURE = False
-    # TODO remove after Svelte frontend is gone
-    SESSION_COOKIE_SAMESITE = "Lax"
-
-    # Admins for local logging
-    ADMINS = [["Local user", "user@localhost"]]
 
     # Show preview of all email types
     PREMAIL_PREVIEW = True
@@ -110,18 +109,20 @@ class Development(Base):
     TEMPLATES = Base.TEMPLATES
     TEMPLATES[0]["OPTIONS"] = {**TEMPLATES[0]["OPTIONS"], "debug": True}
 
+    # Development logging settings
+    # ============================
+    # Simulate sending error emails
+    # Set DEBUG = False to test the admin email handler
+    ADMINS = [["Local user", "user@localhost"]]
     # Logging with timestamps for development
     LOGGING = {
         **Base.LOGGING,
         "formatters": {
+            **Base.LOGGING["formatters"],
             "like_gunicorn": {
                 "format": "[%(asctime)s] %(levelname)-s [%(name)s.%(module)s] ~ %(message)s",
                 "datefmt": "%d/%b/%Y %H:%M:%S",
-            }
-        },
-        "loggers": {
-            **Base.LOGGING["loggers"],
-            "django": {"handlers": ["console"], "propagate": False},
+            },
         },
     }